backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==1.10.14
python-dotenv==1.0.1
pytest==8.2.1
